feat(backend): add remaining fields to model
